requirements.txt
streamlit
langchain
langchain-community
langchain-huggingface
langchain-chroma
langchain-groq
chromadb
pypdf
python-dotenv
huggingface_hub
sentence-transformers
transformers
torch
accelerate